Change to reflect the new module 'synthesis' and the installation of the Synplot/Synspec data files.
The synplot/synspec data files are copied to a hidden folder inside
the user directroy. It is called '.s4'.

As the installation is done by using root priviliges, the root has
ownership of the .s4 folder, and because of that, synplot/synspec
cannot write files in this folder. The last lines of the setup.py
folder is to change ownership from root to user.