Create a function to identify the spectral lines on calculated spectra.
This function read the fort.12 and fort.14 files to identify what
are the spectral lines in the wavelength range asked by the user.

I guess this is similar to the IDL script lineid_select.
Probably, this script will not be used anymore and I could remove it
from the folder.