Change how velocity radial correction acts.
The function rvcorr() to correct for the radial velocity was defined
to put the syntetic spectrum into another frame (contrary what was
described on the docstring). Now, it is defined correctly, i.e,
to put the observation spectrum into the observer frame.

Because this, the function Synplot.apply_rvcorr were deleted.