Fix a bug on the stored value array.
It there were more than one parameter to be fitted, the number of tuples
inside each cell of the stored value array were wrong. It was number of
parameters plus one.

The confusion was that I am working with dtypes and so I do not need
columns on the array. I was setting to the number of parameters plus one
(the +1 was to store the chi square). Now the number of columns is just
one. The store value array is a vector of tuples, in which each tuple
has three values.