Creates a fitting package
This package contains a routine to automatically fit the spectral lines.
The routine was strongly influenced by Ivan Hubeny's `synfit`. In short,
it synthesize spectra in a set of values and compare with the
observation by calculating the chi-square. The minimum chi-square will
give the best values.

There is still room to improve, like applying a convergence method.