Clarify length of windows
I found that the window can only be a single block, i.e., I cannot have
multiple windows.

I added a description about this on the documentation and make it sure
in the code.

It is desirable to have multiple windows, then one can measure, for
example, just the wings of the spectral line.