Fix plotting of chisquare when abund was parameter
When one of the parameter was a chemical element, the plot returned and
error. It was raising the error that it had 2 or more elements, but it
was not. It was catching a wrong error. The problem was with converting
a string from the value of abundance.

I think that it is still catching the error in a wrong way.