Replace code for a more general
This is regarding the plot_chisquare function

I have written this code to be replace the one when finding the best
fit, but the result was the same as it is now (when using a linear
interpolation). Since, in the base, this code was derived from the
previous one on the plot_chisquare function, I decided to replace it.

I hope I did not make it more difficult to understand and then going
against the Zen of Python.