Adds support to no log-scale
The two paameter chi-square plot was always plotted in log-scale. I
added it as optional.

I also changed the way it calculates the log.
Instead of calculating the log of chi-square, I added it as an argument
to the pyplot.imshow command and let matplotlib do the calculation. I
checked with one plot and the plots looked alike.