Explicit the number of parameters when plotting
In the function plot_chisquare, I created a variable that stores the
number of fitted parameters. I hope this will make it more clear.

I also removed a try..except sentence that caught the error if there was
not just 2 parameters. It was not needed because I make sure that there
are two with a if sentence some lines before.